Public interface

InvestmentModel

An AbstractInvestmentModel is a subtype of an EnergyModel as declared in EnergyModelsBase This new type allows for the introduction of investment options to the energy system model.

The composite type InvestmentModel contains some key information for the model such as the emissions limits and penalties for each ResourceEmit, as well as the ResourceEmit instance of CO₂. It uses hence the same fields as an OperationalModel as declared in EnergyModelsBase. In addition, it takes as input the discount_rate. The discount_rate is an important element of investment analysis needed to represent the present value of future cash flows. It is provided to the model as a value between 0 and 1 (e.g. a discount rate of 5 % is 0.05).

EnergyModelsInvestments.AbstractInvestmentModel — Type

An abstract investment model type.

This abstract model type should be used when creating additional EnergyModel types that should utilize investments. An example for additional types is given by the inclusion of, e.g., SDDP.

EnergyModelsInvestments.InvestmentModel — Type

A concrete basic investment model type based on the standard OperationalModel as declared in EnergyModelsBase. The concrete basic investment model is similar to an OperationalModel, but allows for investments and additional discounting of future years.

Fields

emission_limit::Dict{<:ResourceEmit, <:TimeProfile} are the emission caps for the different emissions types considered.

emission_price::Dict{<:ResourceEmit, <:TimeProfile} are the prices for the different emissions types considered.

co2_instance is a ResourceEmit and corresponds to the type used for CO₂.

r::Float64 is the discount rate in the investment optimization.

Additional Data for Investments

Additional data for investment is specified when creating the nodes through subtypes of the type InvestmentData and AbstractInvData. AbstractInvData is introduced in this version to allow for making EnergyModelsInvestments independent of EnergyModelsBase, although this is not yet implemented in the model.

EnergyModelsInvestments.InvestmentData — Type

Abstract type for the extra data for investing in technologies.

EnergyModelsInvestments.AbstractInvData — Type

AbstractInvData

Supertype for investment data for nodal investments.

This additional data is node specific, not technology specific. It is hence possible to provide different values for the same technology through different instances of said technology.

Two types are used to define the parameters necessary for production technologies and transmission modes (SingleInvData) and storages (StorageInvData). The different types are required as the required parameters differ. It is also possible for the user to define new subtypes of InvestmentData if they require additional parameters for including investments in capacities of technologies.

AbstractInvData subtypes

AbstractInvData is used to add the required investment data to the individual technology capacities. The subtypes of AbstractInvData are used for all technologies, that is nodes and transmission modes.

The following fields have to be added for all provided types:

capex::TimeProfile: Capital expenditures (CAPEX) of the Node. The capital expenditures are relative to capacity. Hence, it is important to consider the unit for both costs and the energy of the technology. The total contribution to the objective function $y$ is then given through the equation $y = \texttt{capex} \times x$ where $x$ corresponds to the invested capacity.

max_inst::TimeProfile: Maximum installed capacity of the Node. The maximum installed capacity is limiting the total installed capacity of the Node. It is possible to have different values for different Nodes representing the same technology. This can be useful for, e.g., the potential for wind power in different regions.

inv_mode::Investment: Investment mode of the Node. The individual investment modes are explained in detail in Investment types.

life_mode::LifetimeMode: Lifetime mode of the Node. The lifetime mode is describing how the lifetime of the node is implemented. This includes as well final values and retiring of the individual technologies. The default value is UnlimitedLife. More information can be found in LifetimeMode.

The type StartInvData allows in addition for providing the initial capacity in the first year through:

initial::Real: Starting capacity of the technology in the first strategic period. The starting capacity is only valid for the first strategic period. This capacity will remain present in the simulation horizon, except if retiring is desired by the model. It is not possible to provide a reducing capacity over time for the initial capacity.

while it utilizes the capacity of the technology if the value is not provided through the function EMI.start_cap.

Warning

If you do not use StartInvData, you have to provide the function EMI.start_cap for your type. Otherwise, EnergyModelsInvestments is not able to deduce the starting capcity.

AbstractInvData types have constructors that allow omitting the last field, life_mode.

Warning

All fields that are provided as TimeProfile are accessed in strategic periods. This implies that the provided TimeProfiles are not allowed to have any variations below strategic periods through, e.g. the application of OperationalProfile or RepresentativeProfile.

EnergyModelsInvestments.NoStartInvData — Type

NoStartInvData <: AbstractInvData

Investment data in which the initial capacity is not specified in the InvestmentData. Instead, the initial capacity is inferred from the capacity of the technology.

Fields

capex::TimeProfile is the capital costs for investing in a capacity. The value is relative to the added capacity.

max_inst::TimeProfile is the maximum installed capacity in a strategic period.

inv_mode::Investment is the chosen investment mode for the technology. The following investment modes are currently available: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

life_mode::LifetimeMode is type of handling the lifetime. Several different alternatives can be used: UnlimitedLife, StudyLife, PeriodLife or RollingLife.

EnergyModelsInvestments.StartInvData — Type

StartInvData <: AbstractInvData

Investment data in which the initial capacity is specified in the InvestmentData. The structure is similiar to NoStartInvData with the addition of the field initial::Real, see below.

Fields in addition to NoStartInvData

initial::Real is the initial capacity.

InvestmentData subtypes

InvestmentData subtypes are used to provide technologies introduced in EnergyModelsX (nodes and transmission modes) a subtype of Data that can be used for dispatching. Two different types are directly introduced, SingleInvData and StorageInvData.

SingleInvData is providing a composite type with a single field. It is used for investments in technologies with a single capacity, that is all nodes except for storage nodes as well as tranmission modes.

StorageInvData is required as Storage nodes behave differently compared to the other nodes. In Storage nodes, it is possible to invest both in the charge capacity for storing energy, the storage capacity, that is the level of a Storage node, as well as the discharge capacity, that is how fast energy can be withdrawn. Correspondingly, it is necessary to have individual parameters for the potential investments in each capacity, that is through the fields :charge, :level, and :discharge.

EnergyModelsInvestments.SingleInvData — Type

SingleInvData <: InvestmentData

Extra investment data for type investments. The extra investment data has only a single field in which AbstractInvData has to be added.

The advantage of separating AbstractInvData from the InvestmentData node is to allow easier separation of EnergyModelsInvestments and EnergyModelsBase and provides the user with the potential of introducing new capacities for types.

Fields

cap::AbstractInvData is the investment data for the capacity.

When multiple inputs are provided, a constructor directly creates the corresponding AbstractInvData.

Fields

capex::TimeProfile is the capital costs for investing in a capacity. The value is relative to the added capacity.

max_inst::TimeProfile is the maximum installed capacity in a strategic period.

initial::Real is the initial capacity. This results in the creation of a SingleInvData type for the investment data.

inv_mode::Investment is the chosen investment mode for the technology. The following investment modes are currently available: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

life_mode::LifetimeMode is type of handling the lifetime. Several different alternatives can be used: UnlimitedLife, StudyLife, PeriodLife or RollingLife. If life_mode is not specified, the model assumes an UnlimitedLife.

EnergyModelsInvestments.StorageInvData — Type

StorageInvData <: InvestmentData

Extra investment data for storage investments. The extra investment data for storage investments can, but does not require investment data for the charge capacity of the storage (charge), increasing the storage capacity (level), or the doscharge capacity of the storage (discharge).

It uses the macro @kwdef to use keyword arguments and default values. Hence, the names of the parameters have to be specified.

Fields

charge::Union{AbstractInvData, Nothing} is the investment data for the charge capacity.

level::Union{AbstractInvData, Nothing} is the investment data for the level capacity.

discharge::Union{AbstractInvData, Nothing} is the investment data for the discharge capacity.

Legacy constructors

We provide a legacy constructor, InvData, InvDataStorage, and InvDataTrans, that use the same input as in version 0.5.x. If you want to adjust your model to the latest changes, please refer to the section Update your model to the latest version.

EnergyModelsInvestments.InvData — Function

InvData(;
    capex_cap::TimeProfile,
    cap_max_inst::TimeProfile,
    cap_max_add::TimeProfile,
    cap_min_add::TimeProfile,
    inv_mode::Investment = ContinuousInvestment(),
    cap_start::Union{Real, Nothing} = nothing,
    cap_increment::TimeProfile = FixedProfile(0),
    life_mode::LifetimeMode = UnlimitedLife(),
    lifetime::TimeProfile = FixedProfile(0),
)

Legacy constructor for a InvData.

The new storage descriptions allows now for a reduction in functions which is used to make EnergModelsInvestments less dependent on EnergyModelsBase.

The core changes to the existing structure is the move of the required parameters to the type Investment (e.g., the minimum and maximum added capacity is only required for investment mdodes that require these parameters) as well as moving the lifetime to the type [LifetimeMode], when required..

See the documentation for further information regarding how you can translate your existing model to the new model.

EnergyModelsInvestments.InvDataStorage — Function

InvDataStorage(;
    #Investment data related to storage power
    capex_rate::TimeProfile,
    rate_max_inst::TimeProfile,
    rate_max_add::TimeProfile,
    rate_min_add::TimeProfile,
    capex_stor::TimeProfile,
    stor_max_inst::TimeProfile,
    stor_max_add::TimeProfile,
    stor_min_add::TimeProfile,
    inv_mode::Investment = ContinuousInvestment(),
    rate_start::Union{Real, Nothing} = nothing,
    stor_start::Union{Real, Nothing} = nothing,
    rate_increment::TimeProfile = FixedProfile(0),
    stor_increment::TimeProfile = FixedProfile(0),
    life_mode::LifetimeMode = UnlimitedLife(),
    lifetime::TimeProfile = FixedProfile(0),
)

Storage descriptions were changed in EnergyModelsBase v0.7 resulting in the requirement for rewriting the investment options for Storage nodes.

See the documentation for further information regarding how you can translate your existing model to the new model.

EnergyModelsInvestments.TransInvData — Function

InvData(; capextrans::TimeProfile, transmaxinst::TimeProfile, transmaxadd::TimeProfile, transminadd::TimeProfile, invmode::Investment = ContinuousInvestment(), transstart::Union{Real, Nothing} = nothing, transincrement::TimeProfile = FixedProfile(0), capextransoffset::TimeProfile = FixedProfile(0), )

Legacy constructor for a InvData.

The new storage descriptions allows now for a reduction in functions which is used to make EnergModelsInvestments less dependent on EnergyModelsBase.

The core changes to the existing structure is the move of the required parameters to the type Investment (e.g., the minimum and maximum added capacity is only required for investment mdodes that require these parameters) as well as moving the lifetime to the type [LifetimeMode], when required.

See the documentation for further information regarding how you can translate your existing model to the new model.

Investment modes

Different investment modes are available to help represent different situations. The investment mode is included in the field inv_mode in NoStartInvData and StartInvData. The investment mode how the model can invest and which constraints are imposed on the investments.

Potential fields in investment modes

Investment modes are including the required fields for the investments. These fields are given below with a detailed description in the individual subsections.

max_add::TimeProfile: The maximum added capacity in a strategic period. The maximum added capacity is providing the limit on how fast we can expend a given technology in a strategic period. In general, this value is dependent on the potential construction time and how fast it is possible to build a technology. It is introduced for ContinuousInvestment and SemiContiInvestment modes.

min_add::TimeProfile: The minimum added capacity in a strategic period. The minimum added capacity is providing the lower limit on investments in a strategic period. Its meaning changes, dependent on the chosen investment mode. It is introduced for ContinuousInvestment and SemiContiInvestment modes.

capex_offset::TimeProfile: CAPEX offset for the SemiContinuousOffsetInvestment mode. The offset can be best described with the equation $y = \texttt{capex} \times x + \texttt{capex\_offset}$ where $x$ corresponds to the invested capacity and $y$ to the total capital cost.

cap_increment::TimeProfile: Increment in the case of DiscreteInvestment. The increment corresponds to the potential increase in the case of DiscreteInvestment.

Investment

Investment is the abstract supertype for all investment modes. It is used to allow for a simple extension of the potential investment modes. It is also possible for the user to define new investment modes without major changes to the core structure through specifying a new subtype of Investment.

EnergyModelsInvestments.Investment — Type

Investment

Investment type traits for nodes. The investment type corresponds to the chosen investment mode and includes the required input.

ContinuousInvestment

ContinuousInvestment is the default investment option for all investments, if no alternative is chosen. Continuous investments implies that you can invest in any capacity specified between min_add and max_add. This implies as well that, if min_add is specified, it is necessary to invest in every strategic period in at least this capacity. This approach is the standard approach in large energy system models as it avoids binary variables. However, it may lead to nonsensical solutions, e.g., investments into a 10~MW nuclear power plant.

Warning

Defining min_add::TimeProfile for this mode of investment will lead to a forced investment of at least min_add in each period.

EnergyModelsInvestments.ContinuousInvestment — Type

ContinuousInvestment <: Investment

Continuous investment between a lower and upper bound.

Fields

min_add::TimeProfile is the minimum added capacity in a strategic period. In the case of ContinuousInvestment, this implies that the model must invest at least in this capacity in each strategic period.

max_add::TimeProfile is the maximum added capacity in a strategic period.

BinaryInvestment

BinaryInvestment implies that one can choose to invest to achieve the specified capacity in the given strategic period, or not. The capacity of the investment cannot be adjusted by the optimization.

Warning

This investment type leads to the addition of binary variables. The number of binary variables is equal to the number of strategic periods times the number of Nodes with the BinaryInvestment mode.

EnergyModelsInvestments.BinaryInvestment — Type

BinaryInvestment <: Investment

Binary investment in a given capacity with binary variables. The chosen capacity within a strategic period is given by the field cap.

Binary investments introduce one binary variable for each strategic period.

Fields

cap::TimeProfile is the capacity used for the fixed investments.

DiscreteInvestment

DiscreteInvestment allow for only a discrete increase in the capacity. This increase is specified through the field increment. Hence, it can be also dependent on the strategic period.

DiscreteInvestment can for example be used to represent investment in modular technologies that can be scaled by adding several modules together. In addition, it is beneficial to include for technologies that experience significant economy of scale. In this situation, several instances with different increment and capex can be used

Note

This investment type leads to the addition of integer variables. The number of integer variables is equal to the number of strategic periods times the number of Nodes with the DiscreteInvestment mode.

EnergyModelsInvestments.DiscreteInvestment — Type

DiscreteInvestment <: Investment

Discrete investment with integer variables using an increment. The increment for the discrete investment can be different for the individual strategic periods.

Discrete investments introduce one integer variable for each strategic period.

Fields

increment::TimeProfile is the used increment.

SemiContiInvestment

SemiContiInvestment is an abstract type used for two investment modes:

SemiContinuousInvestment and

SemiContinuousOffsetInvestment.

These investment modes are similar with respect to how you can increase the capacity. They differ however on how the overall cost is calculated. Both investment modes are in general similar to ContinuousInvestment, but the investment is either 0 or between a minimum and maximum value. This means you can define the field min_add::TimeProfile without forcing investment in the technology. Instead, the value determines that if the model decides to invest, then it has to at least invest in the value provided through `minadd`_. This can be also described as:

$x = 0 \lor \texttt{_min\_add} \leq x \leq \texttt{max\_add}$

with $x$ corresponding to the invested capacity

EnergyModelsInvestments.SemiContiInvestment — Type

SemiContiInvestment <: Investment

Supertype for semi-continuous investments, that is the added capacity is either zero or between a minimum and a maximum value.

Semi-continuous investments introduce one binary variable for each strategic period.

Note

These investment modes leads to the addition of binary variables. The number of binary variables is equal to the number of strategic periods times the number of Nodes with the SemiContinuousInvestment and SemiContinuousOffsetInvestment mode.

SemiContinuousInvestment

The cost function in SemiContinuousInvestment is calculated in the same way as in ContinuousInvestment. The total contribution of invested capacity $x$ to the objective function $y$ is given through the equation

$y = \texttt{capex} \times x$.

EnergyModelsInvestments.SemiContinuousInvestment — Type

SemiContinuousInvestment <: Investment

Semi-continuous investments, that is the added capacity is either zero or between a minimum and a maximum value. In this subtype, the cost is crossing the origin, that is the CAPEX is still linear dependent on the

Semi-continuous investments introduce one binary variable for each strategic period.

Fields

min_add::TimeProfile is the minimum added capacity in a strategic period. In the case of SemiContinuousInvestment, this implies that the model must invest at least in this capacity in each strategic period. in this capacity in each strategic period where the model decides to invest. The model can also choose not too invest at all.

max_add::TimeProfile is the maximum added capacity in a strategic period.

SemiContinuousOffsetInvestment

SemiContinuousOffsetInvestment is a type of investment similar to SemiContinuousInvestment and implemented for investments in transmission infrastructure. It does differ with respect to how the costs are calculated. A SemiContinuousOffsetInvestment has an offset in the cost implemented through the the field capex_offset. This offset corresponds to the theoretical cost at an invested capacity of 0.

While SemiContinuousInvestmentutilizes the same relative cost, even if a lower limit is specified, SemiContinuousOffsetInvestment allows for the specification of an offset in the cost through the field capex_offset. This offset is an absolute cost. It corresponds to the theoretical cost at an invested capacity of 0. This changes the contribution to the cost function from

$y = \texttt{capex} \times x$

to

$y = \texttt{capex} \times x + \texttt{capex\_offset}$

where $x$ corresponds to the invested capacity and $y$ to the total capital cost.

EnergyModelsInvestments.SemiContinuousOffsetInvestment — Type

SemiContinuousOffsetInvestment <: Investment

Semi-continuous investments, that is the added capacity is either zero or between a minimum and a maximum value. In this subtype, the cost is not crossing the origin. Instead, there is an offset (y- intercept) in the variable capex_cap, that is its value is larger or smaller than 0 at an invested capacity of 0 given by the field capex_offset. This allows to the user to use different slopes, and hence, account for economy of scales.

Semi-continuous investments introduce one binary variable for each strategic period.

Fields

max_add::TimeProfile is the maximum added capacity in a strategic period.

min_add::TimeProfile is the minimum added capacity in a strategic period. In the case of SemiContinuousOffsetInvestment, this implies that the model must invest at least in this capacity in each strategic period. in this capacity in each strategic period where the model decides to invest. The model can also choose not too invest at all.

capex_offset::TimeProfile is offset for the CAPEX in a strategic period.

FixedInvestment

FixedInvestment is a type of investment where an investment in the given capacity is forced. It allows to account for the investment cost of known investments. In practice, there is however not too much use in including the fixed investment, except if one is interested in the values of the dual variables.

The fields cap_min_add, cap_max_add, and cap_increment do not have a meaning when using FixedInvestment.

EnergyModelsInvestments.FixedInvestment — Type

FixedInvestment <: Investment

Fixed investment in a given capacity. The model is forced to invest in the capacity provided by the field cap.

Fields

cap::TimeProfile is capacity used for the fixed investments.

LifetimeMode

EnergyModelsInvestments allows for differing descriptions for the lifetime of a technology. A key problem is when the lifetime of a technology is not equal to the duration of strategic periods. To this end, several ways to define the lifetime of a technology are available in the package and presented below.

It is also possible for the user to define new LifetimeModes. In practice, this requires only the introduction of a new subtype to LifetimeMode as well as a single function.

EnergyModelsInvestments.LifetimeMode — Type

LifetimeMode

Supertype for the lifetime mode.

UnlimitedLife

This LifetimeMode is used when the lifetime of a Node is not limited. No reinvestment is considered by the optimization and there is also ne salvage value (or rest value) at the end of the last strategic period. Hence, the costs are the same, independent of if the investments in the Node are happening in the first strategic period (and the technology is used for, e.g, 25 years) or the last strategic period (with a usage of, e.g., 5 years) when excluding discounting effects.

UnlimitedLife is the default lifetime mode, if no other mode is specified.

EnergyModelsInvestments.UnlimitedLife — Type

UnlimitedLife <: LifetimeMode

The investment's life is not limited. The investment costs do not consider any reinvestment or rest value.

StudyLife

StudyLife includes the technology for the full investigated horizon. If the Lifetime is shorter than the remaining horizon, reinvestments are considered. These reinvestments are included in the costs of the investment strategic period, but discounted to their actual value.

As an example, consider investments with a lifetime of 20 years in 2030, while the study horizon ends in 2055. In this situation, reinvestments are required in 2050 to allow for operation in the last 5 years. The CAPEX are then correspondingly adjusted to account for both discounted reinvestments and final value in 2055.

EnergyModelsInvestments.StudyLife — Type

StudyLife <: LifetimeMode

The investment lasts for the whole study period with adequate reinvestments at the end of the lifetime and considering the rest value.

Fields

lifetime::TimeProfile is the chosen lifetime of the technology.

PeriodLife

PeriodLife is used to define that the investment is only lasting for the strategic period in which it happens. Additional year of lifetime are counted as a rest value. Reinvestment inside the strategic periods are also considered in case the lifetime is shorter than the length of the strategic period.

EnergyModelsInvestments.PeriodLife — Type

PeriodLife <: LifetimeMode

The investment is considered to last only for the strategic period. The excess lifetime is considered in the rest value. If the lifetime is lower than the length of the period, reinvestment is considered as well.

Fields

lifetime::TimeProfile is the chosen lifetime of the technology.

RollingLife

RollingLife corresponds to the classical roll-over of investments from one strategic period to the next until the end of life is reached. In general, three different cases can be differentiated:

The lifetime is shorter than the duration of the strategic period. In this situation, a PeriodLife is assumed.

The lifetime equals the duration of the strategic period. In this situation, the capacity is retired at the end of the strategic period

The lifetime is longer than the duration of the strategic period. This leaves however a problem if the lifetime does fall in-between two strategic periods, as it would be the case for a lifetime of, e.g., 8 years and two strategic periods of, e.g, 5 years. In this case, the technology would only be available for the first 3 years of the second strategic period leaving the question on how to handle this situation. EnergyModelsInvestments retires the technology at the last full strategic period and calculates the remaining value for the technology.

EnergyModelsInvestments.RollingLife — Type

RollingLife <: LifetimeMode

The investment is rolling to the next strategic periods and it is retired at the end of its lifetime or the end of the previous strategic period if its lifetime ends between two periods.

Fields

lifetime::TimeProfile is the chosen lifetime of the technology.