EnergyModelsInvestments.jl

This Julia package provides investment options for the operational, multi carrier energy model EnergyModelsBase. It furthermore adds investment options for the transmission modes introduced in EnergyModelsGeography, if the package is loaded.

EnergyModelsInvestments — Module

Main module for EnergyModelsInvestments.

This module implements functionalities allowing to run investment analysis.

It is in its current version extending EnergyModelsBase and cannot be used as a stand-alone module.

The extension EMIGeoExt includes furthermore the investment options for transmission modes as described in EnergyModelsGeography.

EnergyModelsInvestments follows the same philosophy with respect to extendibility as EnergyModelsBase. Its aim is to allow the user to come up with new ideas to include investments without major changes to the core structure.

Manual outline

Quick Start

Philosophy

General design philosophy

Investment modes

Lifetime modes

Optimization variables

Cost variables

Capacity variables

Auxiliary variables

Examples

The package is installed with ] add

The code was downloaded with git clone

Library outline

Public interface

InvestmentModel

Additional Data for Investments

Investment modes

LifetimeMode

Internals

Index

Types

Methods