Internals

Index

EMIGeoExt.constraints_transmission_invest

EMIGeoExt.set_capacity_cost

EMIGeoExt.set_trans_cap_installation

EnergyModelsBase.check_node_data

EnergyModelsBase.check_node_data

EnergyModelsBase.constraints_capacity_installed

EnergyModelsBase.constraints_capacity_installed

EnergyModelsBase.objective

EnergyModelsBase.variables_capex

EnergyModelsGeography.update_objective

EnergyModelsGeography.variables_trans_capacity

EnergyModelsGeography.variables_trans_capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.discount_rate

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.investment_data

EnergyModelsInvestments.investment_data

EnergyModelsInvestments.investment_mode

EnergyModelsInvestments.investment_mode

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime_mode

EnergyModelsInvestments.lifetime_mode

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.nodes_investment

EnergyModelsInvestments.nodes_investment

EnergyModelsInvestments.set_capacity_cost

EnergyModelsInvestments.set_capacity_installation

EnergyModelsInvestments.set_capacity_installation

EnergyModelsInvestments.set_capex_discounter

EnergyModelsInvestments.set_investment_properties

EnergyModelsInvestments.start_cap

EnergyModelsInvestments.start_cap

Types

Methods

EnergyModelsBase.check_node_data — Method

EMB.check_node_data(n::EMB.Node, data::InvestmentData, 𝒯, modeltype::AbstractInvestmentModel, check_timeprofiles::Bool)

Performs various checks on investment data for standard nodes.

Checks

Each node can only have a single InvestmentData.

For each field with TimeProfile:

If the TimeProfile is a StrategicProfile, it will check that the profile is in accordance with the TimeStructure

TimeProfiles in InvestmentData cannot include OperationalProfile, RepresentativeProfile, or ScenarioProfile as this is not allowed through indexing on the TimeProfile.

The field :cap_min_add has to be less than :cap_max_add in InvestmentData.

Existing capacity cannot be larger than :cap_max_inst capacity in the beginning. If cap_start is nothing, it also checks that the the field :cap of the node n is not including OperationalProfile, RepresentativeProfile, or ScenarioProfile.

EnergyModelsBase.check_node_data — Method

EMB.check_node_data(
    n::Storage,
    data::InvestmentData,
    𝒯,
    modeltype::AbstractInvestmentModel,
    check_timeprofiles::Bool,
)

Performs various checks on investment data for standard nodes. It is similar to the standard check nodes functions, but adds checks on

Checks

Each node can only have a single InvestmentData.

The InvestmentData must be StorageInvData.

For each individual investment field

For each field with TimeProfile:

If the TimeProfile is a StrategicProfile, it will check that the profile is in accordance with the TimeStructure

TimeProfiles in InvestmentData cannot include OperationalProfile, RepresentativeProfile, or ScenarioProfile as this is not allowed through indexing on the TimeProfile.

The field :min_add has to be less than :max_add.

Existing capacity cannot be larger than :max_inst capacity in the beginning. If NoStartInvData is used, it also checks that the the field :cap of the subfield of node n is not including OperationalProfile, RepresentativeProfile, or ScenarioProfile to avoid indexing problems.

EnergyModelsBase.constraints_capacity_installed — Method

EMB.constraints_capacity_installed(m, n::EMB.Node, 𝒯, modeltype::AbstractInvestmentModel

Set capacity-related constraints for nodes 𝒩 for investment time structure 𝒯:

bounds

binary for BinaryInvestment

link capacity variables

EnergyModelsBase.constraints_capacity_installed — Method

constraints_capacity_installed(m, n::Storage, 𝒯::TimeStructure, modeltype::AbstractInvestmentModel)

Set storage-related constraints for nodes 𝒩ˢᵗᵒʳ for investment time structure 𝒯:

bounds

binary for BinaryInvestment

link storage variables

EnergyModelsBase.objective — Method

EMB.objective(m, 𝒩, 𝒯, modeltype::AbstractInvestmentModel)

Create objective function overloading the default from EMB for AbstractInvestmentModel.

Maximize Net Present Value from investments (CAPEX) and operations (OPEX and emission costs)

TODO:

Consider adding contributions from

revenue (as positive variable, adding positive)

maintenance based on usage (as positive variable, adding negative)

These variables would need to be introduced through the package SparsVariables.

Both are not necessary, as it is possible to include them through the OPEX values, but it would be beneficial for a better separation and simpler calculations from the results.

EnergyModelsBase.variables_capex — Method

EMB.variables_capex(m, 𝒩, 𝒯, 𝒫, modeltype::AbstractInvestmentModel)

Create variables for the capital costs for the invesments in storage and technology nodes.

Additional variables for investment in capacity:

:capex_cap - CAPEX costs for a technology

:cap_current - installed capacity for storage in each strategic period

:cap_add - added capacity

:cap_rem - removed capacity

:cap_invest_b - binary variable whether investments in capacity are happening

:cap_remove_b - binary variable whether investments in capacity are removed

Additional variables for investment in storage:

:stor_level_capex - CAPEX costs for increases in the capacity of a storage

:stor_level_current - installed capacity for storage in each strategic period

:stor_level_add - added capacity

:stor_level_rem - removed capacity

:stor_level_invest_b - binary variable whether investments in capacity are happening

:stor_level_remove_b - binary variable whether investments in capacity are removed

:stor_charge_capex - CAPEX costs for increases in the rate of a storage

:stor_charge_current - installed rate for storage in each strategic period

:stor_charge_add - added rate

:stor_charge_rem - removed rate

:stor_charge_invest_b - binary variable whether investments in rate are happening

:stor_charge_remove_b - binary variable whether investments in rate are removed

EnergyModelsInvestments.capex — Method

capex(n::EMB.Node, t_inv)

Returns the CAPEX of a node n in investment period t_inv.

EnergyModelsInvestments.capex — Method

capex(n::EMB.Node)

Returns the CAPEX of a node n as TimeProfile.

EnergyModelsInvestments.capex — Method

capex(n::Storage, t_inv)

Returns the CAPEX of a Storage node n in investment period t_inv.

EnergyModelsInvestments.capex — Method

capex(n::Storage)

Returns the CAPEX of a Storage node n as TimeProfile.

EnergyModelsInvestments.capex — Method

capex(n::GeneralInvData, t_inv)

Returns the CAPEX of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.capex — Method

capex(inv_data::GeneralInvData)

Returns the CAPEX of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.capex_offset — Method

capex_offset(n::Node, t_inv)

Returns the offset of the CAPEX of node n in investment period t_inv.

EnergyModelsInvestments.discount_rate — Method

discount_rate(modeltype::AbstractInvestmentModel)

Returns the discount rate of EnergyModel modeltype

EnergyModelsInvestments.has_investment — Method

has_investment(n::EMB.Node)

For a given Node, checks that it contains the required investment data.

EnergyModelsInvestments.has_investment — Method

has_investment(n::Storage, field::Symbol)

For a given Storage node, checks that it contains investments for the field field, that is :charge, :level, or :discharge.

EnergyModelsInvestments.increment — Method

increment(n::EMB.Node, t_inv)

Returns the capacity increment of node n in investment period t_inv.

EnergyModelsInvestments.increment — Method

increment(n::EMB.Node)

Returns the capacity increment of node n as TimeProfile.

EnergyModelsInvestments.increment — Method

increment(n::Storage, t_inv)

Returns the capacity increment of Storage node n in investment period t_inv.

EnergyModelsInvestments.increment — Method

increment(n::Storage)

Returns the capacity increment of Storage node n as TimeProfile.

EnergyModelsInvestments.increment — Method

increment(inv_data::GeneralInvData, t_inv)

Returns the capacity increment of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.increment — Method

increment(inv_data::GeneralInvData)

Returns the capacity increment of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.investment_data — Method

investment_data(type, field::Symbol)

Return the investment data of the type type.

EnergyModelsInvestments.investment_data — Method

investment_data(type)

Return the investment data of the type type.

EnergyModelsInvestments.investment_mode — Method

investment_mode(type)

Return the investment mode of the type type. By default, all investments are continuous.

EnergyModelsInvestments.investment_mode — Method

investment_mode(inv_data::GeneralInvData)

Return the investment mode of the investment data inv_data. By default, all investments are continuous.

EnergyModelsInvestments.lifetime — Method

lifetime(type, t)

Return the lifetime of the type type in period t.

EnergyModelsInvestments.lifetime — Method

lifetime(type)

Return the lifetime of the type type as TimeProfile.

EnergyModelsInvestments.lifetime — Method

lifetime(inv_data::GeneralInvData, t)

Return the lifetime of the investment data inv_data in period t.

EnergyModelsInvestments.lifetime — Method

lifetime(inv_data::GeneralInvData)

Return the lifetime of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.lifetime_mode — Method

lifetime_mode(type)

Return the lifetime mode of the type type. By default, all investments are unlimited.

EnergyModelsInvestments.lifetime_mode — Method

lifetime_mode(inv_data::GeneralInvData)

Return the lifetime mode of the investment data inv_data. By default, all investments are unlimited.

EnergyModelsInvestments.max_add — Method

max_add(n::EMB.Node, t_inv)

Returns the maximum allowed added capacity of Node n in investment period t_inv.

EnergyModelsInvestments.max_add — Method

max_add(n::EMB.Node)

Returns the maximum allowed added capacity of Node n as TimeProfile.

EnergyModelsInvestments.max_add — Method

max_add(n::Storage, t_inv)

Returns the maximum allowed added capacity of Storage node n in investment period t_inv.

EnergyModelsInvestments.max_add — Method

max_add(n::Storage)

Returns the maximum allowed added capacity of Storage node n as TimeProfile.

EnergyModelsInvestments.max_add — Method

max_add(inv_data::GeneralInvData, t_inv)

Returns the maximum allowed added capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.max_add — Method

max_add(inv_data::GeneralInvData)

Returns the maximum allowed added capacity of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.max_installed — Method

max_installed(n::EMB.Node, t_inv)

Returns the maximum allowed installed capacity of node n in investment period t_inv.

EnergyModelsInvestments.max_installed — Method

max_installed(n::EMB.Node)

Returns the maximum allowed installed capacity of node n as TimeProfile.

EnergyModelsInvestments.max_installed — Method

max_installed(n::Storage, t_inv)

Returns the maximum allowed installed capacity of Storage node n in investment period t_inv.

EnergyModelsInvestments.max_installed — Method

max_installed(n::Storage)

Returns the maximum allowed installed capacity of Storage node n as TimeProfile.

EnergyModelsInvestments.max_installed — Method

max_installed(inv_data::GeneralInvData, t_inv)

Returns the maximum allowed installed capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.max_installed — Method

max_installed(inv_data::GeneralInvData)

Returns the maximum allowed installed capacity the investment data inv_data as TimeProfile.

EnergyModelsInvestments.min_add — Method

min_add(n::EMB.Node, t_inv)

Returns the minimum allowed added capacity of node n in investment period t_inv.

EnergyModelsInvestments.min_add — Method

min_add(n::EMB.Node)

Returns the minimum allowed added capacity of node n as TimeProfile.

EnergyModelsInvestments.min_add — Method

min_add(n::Storage, t_inv)

Returns the minimum allowed added capacity of Storage node n in investment period t_inv.

EnergyModelsInvestments.min_add — Method

min_add(n::Storage)

Returns the minimum allowed added capacity of Storage node n as TimeProfile.

EnergyModelsInvestments.min_add — Method

min_add(inv_data::GeneralInvData, t_inv)

Returns the minimum allowed added capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.min_add — Method

min_add(inv_data::GeneralInvData)

Returns the minimum allowed added capacity of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.nodes_investment — Method

nodes_investment(𝒩::Vector{<:EMB.Node})

For a given Vector{<:Node}, return all Nodes with investments.

EnergyModelsInvestments.set_capacity_cost — Method

set_capacity_cost(m, n, 𝒯, t_inv, modeltype)

Set the capex_cost based on the technology investment cost, and strategic period length to include the needs for reinvestments and the rest value. It implements different versions of the lifetime implementation:

UnlimitedLife: The investment life is not limited. The investment costs do not consider any reinvestment or rest value.

StudyLife: The investment last for the whole study period with adequate reinvestments at end of lifetime and rest value.

PeriodLife: The investment is considered to last only for the strategic period. the excess lifetime is considered in the rest value.

RollingLife: The investment is rolling to the next strategic periods and it is retired at the end of its lifetime or the end of the previous sp if its lifetime ends between two sp.

EnergyModelsInvestments.set_capacity_installation — Method

set_storage_installation(m, n, 𝒯ᴵⁿᵛ)

Add constraints related to storage installation depending on investment mode of node n

EnergyModelsInvestments.set_capacity_installation — Method

set_capacity_installation(m, n, 𝒯ᴵⁿᵛ)

Add constraints related to capacity installation depending on investment mode of node n

EnergyModelsInvestments.set_capex_discounter — Method

set_capex_discounter(years, lifetime, r)

Calculate the discounted values used in the lifetime calculations, when the LifetimeMode is given by PeriodLife and StudyLife.

Arguments

years:: the remaining years for calculating the discounted value. The years are

depending on the considered LifetimeMode, using remaining(t_inv, 𝒯) for StudyLife and duration(t_inv) for PeriodLife.

lifetime: the lifetime of the node.

r: the discount rate.

EnergyModelsInvestments.set_investment_properties — Method

set_investment_properties(n, var)

Set investment properties for variable var for type n, e.g., set to binary for BinaryInvestment, bounds, etc.

EnergyModelsInvestments.start_cap — Method

start_cap(m, n, t, stcap, modeltype)

Returns the starting capacity of the node in the first investment period. If no starting capacity is provided in InvestmentData (default = Nothing), then use the provided capacity from the field Cap.

EMIGeoExt.constraints_transmission_invest — Method

constraints_transmission_invest(m, 𝒯, ℳ, modeltype::EMI.AbstractInvestmentModel)

Set capacity-related constraints for TransmissionModes ℳ for investment time structure 𝒯:

bounds

binary for BinaryInvestment

link capacity variables

EMIGeoExt.set_capacity_cost — Method

set_capacity_cost(m, tm::TransmissionMode, 𝒯, t_inv, modeltype)

Set capex_trans based on the technology investment cost to include the potential for either semi continuous costs with offsets or piecewise linear costs. It implements different versions of cost calculations:

Investment: The cost is linear dependent on the installed capacity. This is the default

for all invcestment options

SemiContinuousOffsetInvestment: The cost is linear dependent on the added capacity with a

given offset

EMIGeoExt.set_trans_cap_installation — Method

set_trans_cap_installation(m, tm, 𝒯ᴵⁿᵛ, investment_mode)

Add constraints related to capacity installation depending on investment mode of TransmissionMode tm.

EnergyModelsGeography.update_objective — Method

EMG.update_objective(m, 𝒩, 𝒯, 𝒫, ℒᵗʳᵃⁿˢ, modeltype::EMI.AbstractInvestmentModel)

Create objective function overloading the default from EMB for EMI.AbstractInvestmentModel.

Maximize Net Present Value from revenues, investments (CAPEX) and operations (OPEX)

TODO:

* consider passing expression around for updating

* consider reading objective and adding terms/coefficients (from model object m)

EnergyModelsGeography.variables_trans_capacity — Method

EMG.variables_trans_capacity(m, 𝒯, ℳ, modeltype::EMI.AbstractInvestmentModel)

Create variables to track how much of installed transmision capacity is used for all time periods t ∈ 𝒯 and how much energy is lossed. Introduction of the additional constraints for investments.

Additional variables for investment in capacity:

:trans_cap_invest_b - binary variable whether investments in capacity are happening

:trans_cap_remove_b - binary variable whether investments in capacity are removed

:trans_cap_current - installed capacity for storage in each strategic period

:trans_cap_add - added capacity

:trans_cap_rem - removed capacity

EnergyModelsGeography.variables_trans_capex — Method

EMG.variables_trans_capex(m, 𝒯, ℳ,, modeltype::EMI.AbstractInvestmentModel)

Create variables for the capital costs for the investments in transmission.

EnergyModelsInvestments.capex — Method

EMI.capex(n::TransmissionMode, t_inv)

Returns the CAPEX of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.capex — Method

EMI.capex(n::TransmissionMode)

Returns the CAPEX of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.capex_offset — Method

capex_offset(n::TransmissionMode, t_inv)

Returns the offset of the CAPEX of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.capex_offset — Method

capex_offset(n::TransmissionMode)

Returns the offset of the CAPEX of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.has_investment — Method

EMI.has_investment(tm::TransmissionMode)

For a given TransmissionMode, checks that it contains ithe required investment data.

EnergyModelsInvestments.increment — Method

EMI.increment(tm::TransmissionMode, t_inv)

Returns the minimum added capacity of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.increment — Method

EMI.increment(tm::TransmissionMode)

Returns the minimum added capacity of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.max_add — Method

EMI.max_add(n::TransmissionMode, t_inv)

Returns the maximum allowed added capacity of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.max_add — Method

EMI.max_add(n::TransmissionMode)

Returns the maximum allowed added capacity of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.max_installed — Method

EMI.max_installed(n::TransmissionMode, t_inv)

Returns the maximum allowed installed capacity of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.max_installed — Method

EMI.max_installed(n::TransmissionMode)

Returns the maximum allowed installed capacity of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.min_add — Method

EMI.min_add(n::TransmissionMode, t_inv)

Returns the minimum allowed added capacity of TransmissionMode tm in investment period t_inv.

EnergyModelsInvestments.min_add — Method

EMI.min_add(n::TransmissionMode)

Returns the minimum allowed added capacity of TransmissionMode tm as TimeProfile.

EnergyModelsInvestments.nodes_investment — Method

EMI.has_investment(ℳ::Vector{<:TransmissionMode})

For a given Vector{<:TransmissionMode}, return all TransmissionModes with investments.

EnergyModelsInvestments.start_cap — Method

EMI.start_cap(m, tm, t, stcap, modeltype)

Returns the starting capacity of the TransmissionMode tm in the first investment period. If no starting capacity is provided in InvestmentData (default = Nothing), then use the provided capacity from the field trans_Cap.