Internals

Index

EnergyModelsBase.check_node_data

EnergyModelsBase.check_node_data

EnergyModelsBase.constraints_capacity_installed

EnergyModelsBase.constraints_capacity_installed

EnergyModelsBase.objective

EnergyModelsBase.variables_capex

EnergyModelsGeography.constraints_capacity_installed

EnergyModelsGeography.update_objective

EnergyModelsGeography.variables_trans_capex

EnergyModelsInvestments.add_investment_constraints

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.capex_offset

EnergyModelsInvestments.check_inv_data

EnergyModelsInvestments.discount_rate

EnergyModelsInvestments.get_var_add

EnergyModelsInvestments.get_var_add

EnergyModelsInvestments.get_var_capex

EnergyModelsInvestments.get_var_capex

EnergyModelsInvestments.get_var_current

EnergyModelsInvestments.get_var_current

EnergyModelsInvestments.get_var_inst

EnergyModelsInvestments.get_var_inst

EnergyModelsInvestments.get_var_inst

EnergyModelsInvestments.get_var_invest_b

EnergyModelsInvestments.get_var_invest_b

EnergyModelsInvestments.get_var_rem

EnergyModelsInvestments.get_var_rem

EnergyModelsInvestments.get_var_remove_b

EnergyModelsInvestments.get_var_remove_b

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.increment

EnergyModelsInvestments.invest_capacity

EnergyModelsInvestments.invest_capacity

EnergyModelsInvestments.invest_capacity

EnergyModelsInvestments.invest_capacity

EnergyModelsInvestments.investment_data

EnergyModelsInvestments.investment_data

EnergyModelsInvestments.investment_data

EnergyModelsInvestments.investment_mode

EnergyModelsInvestments.investment_mode

EnergyModelsInvestments.investment_mode

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime

EnergyModelsInvestments.lifetime_mode

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_add

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.max_installed

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.min_add

EnergyModelsInvestments.nodes_investment

EnergyModelsInvestments.set_capacity_cost

EnergyModelsInvestments.set_capacity_installation

EnergyModelsInvestments.set_capex_discounter

EnergyModelsInvestments.set_capex_value

EnergyModelsInvestments.set_capex_value

EnergyModelsInvestments.set_capex_value

EnergyModelsInvestments.start_cap

Types

Methods

EnergyModelsBase.check_node_data — Method

EMB.check_node_data(
    n::EMB.Node,
    data::InvestmentData,
    𝒯,
    modeltype::AbstractInvestmentModel,
    check_timeprofiles::Bool
)

Performs various checks on investment data for standard nodes.

Checks

Each node can only have a single InvestmentData.

All checks incorporated in the function check_inv_data.

EnergyModelsBase.check_node_data — Method

EMB.check_node_data(
    n::Storage,
    data::InvestmentData,
    𝒯,
    modeltype::AbstractInvestmentModel,
    check_timeprofiles::Bool,
)

Performs various checks on investment data for standard nodes. It is similar to the standard check nodes functions, but adds checks on

Checks

Each node can only have a single InvestmentData.

The InvestmentData must be StorageInvData.

For each individual investment field all checks incorporated in the function check_inv_data.

EnergyModelsBase.constraints_capacity_installed — Method

EMB.constraints_capacity_installed(
    m,
    n::EMB.Node,
    𝒯::TimeStructure,
    modeltype::AbstractInvestmentModel,
)

When the modeltype is an investment model, the function introduces the related constraints for the capacity expansion. The investment mode and lifetime mode are used for adding constraints.

The default function only accepts nodes with SingleInvData. If you have several capacities for investments, you have to dispatch specifically on the node type. This is implemented for Storage nodes.

EnergyModelsBase.constraints_capacity_installed — Method

EMB.constraints_capacity_installed(
    m,
    n::Storages,
    𝒯::TimeStructure,
    modeltype::AbstractInvestmentModel,
)

When the modeltype is an investment model and the node is a Storage node, the function introduces the related constraints for the capacity expansions for the fields :charge, :level, and :discharge. This requires the utilization of the StorageInvData investment type, in which the investment mode and lifetime mode are used for adding constraints for each capacity.

EnergyModelsBase.objective — Method

EMB.objective(m, 𝒩, 𝒯, 𝒫, modeltype::AbstractInvestmentModel)

Create objective function overloading the default from EMB for AbstractInvestmentModel.

Maximize Net Present Value from investments (CAPEX) and operations (OPEX and emission costs)

TODO:

Consider adding contributions from

revenue (as positive variable, adding positive)

maintenance based on usage (as positive variable, adding negative)

These variables would need to be introduced through the package SparsVariables.

Both are not necessary, as it is possible to include them through the OPEX values, but it would be beneficial for a better separation and simpler calculations from the results.

EnergyModelsBase.variables_capex — Method

EMB.variables_capex(m, 𝒩, 𝒯, 𝒫, modeltype::AbstractInvestmentModel)

Create variables for the capital costs for the invesments in storage and technology nodes.

Additional variables for investment in capacity:

:cap_capex - CAPEX costs for a technology

:cap_current - installed capacity for storage in each strategic period

:cap_add - added capacity

:cap_rem - removed capacity

:cap_invest_b - binary variable whether investments in capacity are happening

:cap_remove_b - binary variable whether investments in capacity are removed

Additional variables for investment in storage:

:stor_level_capex - CAPEX costs for increases in the capacity of a storage

:stor_level_current - installed capacity for storage in each strategic period

:stor_level_add - added capacity

:stor_level_rem - removed capacity

:stor_level_invest_b - binary variable whether investments in capacity are happening

:stor_level_remove_b - binary variable whether investments in capacity are removed

:stor_charge_capex - CAPEX costs for increases in the rate of a storage

:stor_charge_current - installed rate for storage in each strategic period

:stor_charge_add - added rate

:stor_charge_rem - removed rate

:stor_charge_invest_b - binary variable whether investments in rate are happening

:stor_charge_remove_b - binary variable whether investments in rate are removed

EnergyModelsInvestments.add_investment_constraints — Method

add_investment_constraints(
    m,
    element,
    inv_data::AbstractInvData,
    cap,
    prefix,
    𝒯ᴵⁿᵛ::TS.StratPeriods,
    disc_rate::Float64,
)

Core function for introducing constraints on the investments. The constraints include introducing bounds on the available capacities as well as the calculation of the CAPEX.

The function calls two additional subroutines, set_capacity_installation and set_capacity_cost which are used for introducing bounds on the investment variables and calculating the CAPEX contribution of each investments. The utilization of subroutines allows the introduction of dispatch for the individual investment and lifetime options.

Arguments

m: the JuMP model instance.

element: the element for which investment constraints should be added. Any potential element can be used. In EnergyModelsBase, the individual element is either a Node or a TransmissionMode.

inv_data::AbstractInvData: the investment data for the node and capacity cap.

prefix: the prefix used for variables for this element. This argument is used for extracting the individual investment variables.

cap: the field that is used if several capacities are provided.

𝒯ᴵⁿᵛ::TS.StratPeriods: the strategic periods structure.

disc_rate: the discount rate used in the lifetime calculation for reinvestment and end of life calculations.

EnergyModelsInvestments.capex — Method

capex(n::AbstractInvData, t_inv)

Returns the CAPEX of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.capex — Method

capex(inv_data::AbstractInvData)

Returns the CAPEX of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.capex_offset — Method

capex_offset(inv_data::AbstractInvData, t_inv)

Returns the offset of the CAPEX of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.capex_offset — Method

capex_offset(inv_data::AbstractInvData)

Returns the offset of the CAPEX of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.capex_offset — Method

capex_offset(inv_mode::SemiContinuousOffsetInvestment, t_inv)

Returns the offset of the CAPEX of the investment mode inv_mode in investment period t_inv.

EnergyModelsInvestments.capex_offset — Method

capex_offset(inv_mode::SemiContinuousOffsetInvestment)

Returns the offset of the CAPEX of the investment mode inv_mode as TimeProfile.

EnergyModelsInvestments.check_inv_data — Method

check_inv_data(
    inv_data::AbstractInvData,
    capacity_profile::TimeProfile,
    𝒯,
    message::String,
    check_timeprofiles::Bool,
)

Performs various checks on investment data introduced within EnergyModelsInvestments

Checks

For each field with TimeProfile:

If the TimeProfile is a StrategicProfile, it will check that the profile is in accordance with the TimeStructure

TimeProfiles in InvestmentData cannot include OperationalProfile, RepresentativeProfile, or ScenarioProfile as this is not allowed through indexing on the TimeProfile.

The field :min_add has to be less than :max_add if the investment mode is given by ContinuousInvestment or SemiContiInvestment.

Existing capacity cannot be larger than :max_inst capacity in the beginning. If NoStartInvData is used, it also checks that the the TimeProfile capacity_profile is not including OperationalProfile, RepresentativeProfile, or ScenarioProfile to avoid indexing problems.

EnergyModelsInvestments.discount_rate — Method

discount_rate(modeltype::AbstractInvestmentModel)

Returns the discount rate of EnergyModel modeltype

EnergyModelsInvestments.get_var_add — Method

get_var_add(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_add — Method

get_var_add(m, prefix::Symbol)

Extracts the investment capacity variable with a given prefix from the model.

EnergyModelsInvestments.get_var_capex — Method

get_var_capex(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_capex — Method

get_var_capex(m, prefix::Symbol)

Extracts the CAPEX variable with a given prefix from the model.

EnergyModelsInvestments.get_var_current — Method

get_var_current(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_current — Method

get_var_current(m, prefix::Symbol)

Extracts the current capacity variable with a given prefix from the model.

EnergyModelsInvestments.get_var_inst — Method

get_var_inst(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_inst — Method

get_var_inst(m, prefix::Symbol)

Extracts the installed capacity variable with a given prefix from the model.

EnergyModelsInvestments.get_var_invest_b — Method

get_var_invest_b(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_invest_b — Method

get_var_invest_b(m, prefix::Symbol)

Extracts the binary investment variable with a given prefix from the model.

EnergyModelsInvestments.get_var_rem — Method

get_var_rem(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_rem — Method

get_var_rem(m, prefix::Symbol)

Extracts the retired capacity variable with a given prefix from the model.

EnergyModelsInvestments.get_var_remove_b — Method

get_var_remove_b(m, prefix::Symbol, element)

When the type element is used as conditional input, it extracts only the variable for the specified node.

EnergyModelsInvestments.get_var_remove_b — Method

get_var_remove_b(m, prefix::Symbol)

Extracts the binary retirement variable with a given prefix from the model.

EnergyModelsInvestments.has_investment — Method

has_investment(element)

For a given type element, checks that it contains the required investment data.

EnergyModelsInvestments.has_investment — Method

has_investment(n::Storage, field::Symbol)

When the element type is a Storage node, checks that it contains investments for the field field, that is :charge, :level, or :discharge.

EnergyModelsInvestments.increment — Method

increment(inv_data::AbstractInvData, t_inv)

Returns the capacity increment of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.increment — Method

increment(inv_data::AbstractInvData)

Returns the capacity increment of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.increment — Method

increment(inv_mode::Investment, t_inv)

Returns the capacity increment of the investment mode inv_mode in investment period t_inv.

EnergyModelsInvestments.increment — Method

increment(inv_mode::Investment)

Returns the capacity increment of the investment mode inv_mode as TimeProfile.

EnergyModelsInvestments.invest_capacity — Method

invest_capacity(inv_data::AbstractInvData, t_inv)

Returns the capacity profile for investments of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.invest_capacity — Method

invest_capacity(inv_data::AbstractInvData)

Returns the capacity investments of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.invest_capacity — Method

invest_capacity(inv_mode::Investment, t_inv)

Returns the capacity profile for investments of the investment mode inv_mode in investment period t_inv.

EnergyModelsInvestments.invest_capacity — Method

invest_capacity(inv_mode::Investment)

Returns the capacity investments of the investment mode inv_mode as TimeProfile.

EnergyModelsInvestments.investment_data — Method

investment_data(element, field::Symbol)

Return the investment data of the type element of the capacity field.

EnergyModelsInvestments.investment_data — Method

investment_data(element)

Return the investment data of the type element.

EnergyModelsInvestments.investment_data — Method

investmentdata(invdata::SingleInvData)

Return the investment data of the investment data SingleInvData.

EnergyModelsInvestments.investment_mode — Method

investment_mode(inv_data::AbstractInvData)

Return the investment mode of the investment data inv_data. By default, all investments are continuous.

EnergyModelsInvestments.investment_mode — Method

investment_mode(element, cap::Symbol)

Return the investment mode of the type element and the capacity cap.

This function utilizes the function investment_mode(investment_mode(inv_data::AbstractInvData)) for the AbstractInvData of the capacity cap

EnergyModelsInvestments.investment_mode — Method

investment_mode(element)

Return the investment mode of the type element. By default, all investments are continuous.

EnergyModelsInvestments.lifetime — Method

lifetime(inv_data::AbstractInvData, t_inv)

Return the lifetime of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.lifetime — Method

lifetime(inv_data::AbstractInvData)

Return the lifetime of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.lifetime — Method

lifetime(lifetime_mode::LifetimeMode, t_inv)

Return the lifetime of the lifetime mode lifetime_mode in investment period t_inv.

EnergyModelsInvestments.lifetime — Method

lifetime(lifetime_mode::LifetimeMode)

Return the lifetime of the lifetime mode lifetime_mode as TimeProfile.

EnergyModelsInvestments.lifetime_mode — Method

lifetime_mode(inv_data::AbstractInvData)

Return the lifetime mode of the investment data inv_data. By default, all investments are unlimited.

EnergyModelsInvestments.max_add — Method

max_add(inv_data::AbstractInvData, t_inv)

Returns the maximum allowed added capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.max_add — Method

max_add(inv_data::AbstractInvData)

Returns the maximum allowed added capacity of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.max_add — Method

max_add(inv_mode::Investment, t_inv)

Returns the maximum allowed added capacity of the investment mode inv_mode investment period t_inv.

EnergyModelsInvestments.max_add — Method

max_add(inv_mode::Investment)

Returns the maximum allowed added capacity of the investment mode inv_mode as TimeProfile.

EnergyModelsInvestments.max_installed — Method

max_installed(inv_data::AbstractInvData, t_inv)

Returns the maximum allowed installed capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.max_installed — Method

max_installed(inv_data::AbstractInvData)

Returns the maximum allowed installed capacity the investment data inv_data as TimeProfile.

EnergyModelsInvestments.min_add — Method

min_add(inv_data::AbstractInvData, t_inv)

Returns the minimum allowed added capacity of the investment data inv_data in investment period t_inv.

EnergyModelsInvestments.min_add — Method

min_add(inv_data::AbstractInvData)

Returns the minimum allowed added capacity of the investment data inv_data as TimeProfile.

EnergyModelsInvestments.min_add — Method

min_add(inv_mode::Investment, t_inv)

Returns the minimum allowed added capacity of the investment mode inv_mode in investment period t_inv.

EnergyModelsInvestments.min_add — Method

min_add(inv_mode::Investment)

Returns the minimum allowed added capacity of the investment mode inv_mode as TimeProfile.

EnergyModelsInvestments.nodes_investment — Method

nodes_investment(𝒩::Vector{<:EMB.Node})

For a given Vector{<:Node}, return all Nodes with investments.

EnergyModelsInvestments.set_capacity_cost — Method

set_capacity_cost(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ, disc_rate)

Set the capex_cost based on the technology investment cost, and strategic period length to include the needs for reinvestments and the rest value. It implements different versions of the lifetime implementation:

UnlimitedLife: The investment life is not limited. The investment costs do not consider any reinvestment or rest value.

StudyLife: The investment last for the whole study period with adequate reinvestments at end of lifetime and rest value.

PeriodLife: The investment is considered to last only for the strategic period. the excess lifetime is considered in the rest value.

RollingLife: The investment is rolling to the next strategic periods and it is retired at the end of its lifetime or the end of the previous sp if its lifetime ends between two sp.

EnergyModelsInvestments.set_capacity_installation — Method

set_capacity_installation(m, element, prefix, 𝒯ᴵⁿᵛ, inv_mode)

Add constraints related to installation depending on investment mode of type element.

EnergyModelsInvestments.set_capex_discounter — Method

set_capex_discounter(years, lifetime, disc_rate)

Calculate the discounted values used in the lifetime calculations, when the LifetimeMode is given by PeriodLife and StudyLife.

Arguments

years:: the remaining years for calculating the discounted value. The years are depending on the considered LifetimeMode, using remaining(t_inv, 𝒯) for StudyLife and duration(t_inv) for PeriodLife.

lifetime: the lifetime of the element.

disc_rate: the discount rate.

EnergyModelsInvestments.set_capex_value — Method

set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ)

Calculate the cost value for the different investment modes of the investment data inv_data for element element.

Arguments

m: the JuMP model instance.

element: the element type for which the absolute CAPEX should be calculated.

r: the discount rate.

inv_data: the investment data given as subtype of AbstractInvData.

prefix: the prefix used for variables for this element.

𝒯ᴵⁿᵛ: the strategic periods structure.

EnergyModelsInvestments.set_capex_value — Method

set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ, ::Investment)

When no specialized method is defined for the investment mode, it calculates the capital cost based on the multiplication of the field capex in inv_data with the added capacity.

EnergyModelsInvestments.set_capex_value — Method

set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ, inv_mode::SemiContinuousOffsetInvestment)

When the investment mode is given by SemiContinuousOffsetInvestment then there is an additional offset for the CAPEX.

EnergyModelsInvestments.start_cap — Method

start_cap(element, t_inv, inv_data::AbstractInvData, cap)

Returns the starting capacity of the type element in the first investment period. If NoStartInvData is used for the starting capacity, it deduces the value from the provided initial capacity.

EnergyModelsGeography.constraints_capacity_installed — Method

EMG.constraints_capacity_installed(
    m,
    tm::TransmissionMode,
    𝒯::TimeStructure,
    modeltype::AbstractInvestmentModel,
)

When the modeltype is an investment model, the function introduces the related constraints for the capacity expansion. The investment mode and lifetime mode are used for adding constraints.

The default function only accepts nodes with SingleInvData. If you have several capacities for investments, you have to dispatch specifically on the function. This is implemented for Storage nodes.

EnergyModelsGeography.update_objective — Method

EMG.update_objective(m, 𝒩, 𝒯, 𝒫, ℒᵗʳᵃⁿˢ, modeltype::EMI.AbstractInvestmentModel)

Create objective function overloading the default from EMB for EMI.AbstractInvestmentModel.

Maximize Net Present Value from revenues, investments (CAPEX) and operations (OPEX)

TODO:

* consider passing expression around for updating

* consider reading objective and adding terms/coefficients (from model object m)

EnergyModelsGeography.variables_trans_capex — Method

EMG.variables_trans_capex(m, 𝒯, ℳ,, modeltype::EMI.AbstractInvestmentModel)

Create variables for the capital costs for the investments in transmission.

Additional variables for investment in capacity:

:trans_cap_capex - CAPEX costs for increases in the capacity of a transmission mode

:trans_cap_current - installed capacity for storage in each strategic period

:trans_cap_add - added capacity

:trans_cap_rem - removed capacity

:trans_cap_invest_b - binary variable whether investments in capacity are happening

:trans_cap_remove_b - binary variable whether investments in capacity are removed

EnergyModelsInvestments.get_var_inst — Method

EMI.get_var_inst(m, prefix::Symbol, tm::EMG.TransmissionMode)

When the transmission mode tm is used as conditional input, it extracts only the variable for the specified transmission mode.