Public interface

InvestmentModel

An AbstractInvestmentModel is a subtype of an EnergyModel as declared in EnergyModelsBase This new type allows for the introduction of investment options to the energy system model.

The composite type InvestmentModel contains some key information for the model such as the emissions limits and penalties for each ResourceEmit, as well as the ResourceEmit instance of CO₂. It uses hence the same fields as an OperationalModel as declared in EnergyModelsBase. In addition, it takes as input the discount_rate. The discount_rate is an important element of investment analysis needed to represent the present value of future cash flows. It is provided to the model as a value between 0 and 1 (e.g. a discount rate of 5 % is 0.05).

EnergyModelsInvestments.AbstractInvestmentModel — Type

An abstract investment model type.

This abstract model type should be used when creating additional EnergyModel types that should utilize investments. An example for additional types is given by the inclusion of, e.g., SDDP.

EnergyModelsInvestments.InvestmentModel — Type

A concrete basic investment model type based on the standard OperationalModel as declared in EnergyModelsBase. The concrete basic investment model is similar to an OperationalModel, but allows for investments and additional discounting of future years.

Fields

emission_limit::Dict{<:ResourceEmit, <:TimeProfile} are the emission caps for the different emissions types considered.

emission_price::Dict{<:ResourceEmit, <:TimeProfile} are the prices for the different emissions types considered.

co2_instance is a ResourceEmit and corresponds to the type used for CO₂.

r is the discount rate in the investment optimization.

Additional Data for Investments

Additional data for investment is specified when creating the nodes through subtypes of the type InvestmentData.

EnergyModelsInvestments.InvestmentData — Type

Abstract type for the extra data for investing in technologies.

This additional data is node specific, not technology specific. It is hence possible to provide different values for the same technology through different instances of said technology.

Two types are used to define the parameters necessary for production technologies (InvData) and storages (InvDataStorage), while one is used for transmission modes (TransInvData) through the EMIGeoExt extension. The different types are required as the required parameters differ. It is also possible for the user to define new subtypes of InvestmentData if they require additional parameters for including investments in technologies.

Note

Depending on the chosen investment mode, not all parameters are necessary. It is however possible to set parameters even if they will not be used. This allows for a simple change of the investment mode.

The required parameters will be described in Investment Types for each investment option. All InvestmentData types use Base.@kwdef to simplify the creation of constructors. Hence, it is necessary to specify the the name of the parameters when creating instances of the types.

Note

It is planned in a future iteration to revamp how we provide the investment data. A key change is related to the approach for Storage and TransmissionMode investments to reduce repetitions in the functions.

InvData

InvData is used to add the required investment data to the individual technology nodes. It is used for almost all technology nodes, except for Storage nodes.

The following fields have to be added as keyword arguments:

capex_cap::TimeProfile: Capital expenditures (CAPEX) of the Node. The capital expenditures are relative costs. Hence, it is important to consider the unit for both costs and the energy of the technology. The total contribution to the objective function $y$ is then given through the equation $y = \texttt{capex\_cap} \times x$ where $x$ corresponds to the invested capacity.

cap_max_inst::TimeProfile: Maximum installed capacity of the Node. The maximum installed capacity is limiting the total installed capacity of the Node. It is possible to have different values for different Nodes representing the same technology. This can be useful for, e.g., the potential for wind power in different regions.

cap_max_add::TimeProfile: The maximum added capacity in a strategic period. The maximum added capacity is providing the limit on how fast we can expend a given technology in a strategic period. In general, this value is dependent on the potential construction time and how fast it is possible to build a technology.

cap_min_add::TimeProfile: The minimum added capacity in a strategic period. The minimum added capacity is providing the lower limit on investments in a strategic period. Its meaning changes, dependent on the chosen investment mode, as outlined in Investment types.

In addition, there are also parameters with a default value. These values do not have to be provided, but can be provided, if desired. The following parameter can be modified:

inv_mode::Investment: Investment mode of the Node. The individual investment modes are explained in detail in Investment types. The default investment mode is ContinuousInvestment.

cap_start::Union{Real, Nothing}: Starting capacity of the Node in the first strategic period. The starting capacity is only valid for the first strategic period. This capacity will remain present in the simulation horizon, except if retiring is desired by the model. It is not possible to provide a reducing capacity over time for the initial capacity. The model extracts alternatively the value of the capacity in the first strategic period, if no value is provided. This is the default value.

cap_increment::TimeProfile: Increment in the case of DiscreteInvestment. The increment corresponds to the potential increase in the case of DiscreteInvestment. Its usage will be explained further in Investment types. The default value is 0 corresponding to no investment possibility.

life_mode::LifetimeMode: Lifetime mode of the Node. The lifetime mode is describing how the lifetime of the node is implemented. This includes as well final values and retiring of the individual technologies. The default value is UnlimitedLife. More information can be found in LifetimeMode.

lifetime::TimeProfile: Lifetime of the Node. The lifetime corresponds to the lifetime of the invested Node. Its requirement is dependent on the chosen LifetimeMode. The default value is 0 corresponding to no lifetime.

Warning

All fields that are provided as TimeProfile are accessed in strategic periods. This implies that the provided TimeProfiles are not allowed to have any variations below strategic periods through, e.g. the application of OperationalProfile or RepresentativeProfile.

EnergyModelsInvestments.InvData — Type

Extra data for investing in technologies.

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses the macro Base.@kwdef to use keyword arguments and default values. Hence, the name of the parameters have to be specified.

Fields

capex_cap::TimeProfile Capital expenditure for the capacity in a strategic period.

cap_max_inst::TimeProfile Maximum possible installed capacity of the technology in a strategic period.

cap_max_add::TimeProfile Maximum capacity addition in a strategic period.

cap_min_add::TimeProfile Minimum capacity addition in a strategic period.

inv_mode::Investment = ContinuousInvestment() Type of the investment: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment, or FixedInvestment.

cap_start::Union{Real, Nothing} = nothing Starting capacity in first period. If nothing is given, it is set by start_cap() to the capacity cap of the node in the first strategic period.

cap_increment::TimeProfile = FixedProfile(0) Capacity increment used in the case of DiscreteInvestment.

life_mode::LifetimeMode = UnlimitedLife() Type of handling of the lifetime: UnlimitedLife, StudyLife, PeriodLife or RollingLife

lifetime::TimeProfile = FixedProfile(0) Duration/lifetime of the technology invested in each period.

InvDataStorage

InvDataStorage is required as Storage nodes behave differently compared to the other nodes. In Storage nodes, it is possible to invest both in the rate for storing energy as well as in the storage capacity, that is the level of a Storage node. Correspondingly, it is necessary to have individual parameters for both the rate and the level. Storage nodes have in general the same fields with a slightly different naming to account for both rate and level investments:

capex_cap is provided for both the rate (capex_rate) and the level (capex_stor),

cap_max_inst is provided for both the rate (rate_max_inst) and the level (stor_max_inst),

cap_max_add is provided for both the rate (rate_max_add) and the level (stor_max_add),

cap_min_add is provided for both the rate (rate_min_add) and the level (stor_min_add),

cap_start is provided for both the rate (rate_start) and the level (stor_start), and

cap_increment is provided for both the rate (rate_increment) and the level (stor_increment).

Both the Investment mode and LifetimeMode are the same for investments in rate and storage level. The same holds as well for the lifetime of the technology.

The required fields are the same as in InvData.

EnergyModelsInvestments.InvDataStorage — Type

Extra data for investing in storages.

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses the macro Base.@kwdef to use keyword arguments and default values. Hence, the name of the parameters have to be specified.

The parameters are separated between rate_ and stor_. The rate_ parameters refer to rate components (power, flow, ...) for instance, charging and discharging power of batteries, while the stor_ refers to a volumetric component (energy, volume, mass...), for instance storage capacity of a battery.

Fields

capex_rate::TimeProfile Capital expenditure for storage rate, here investment costs of the technology rate in each period.

rate_max_inst::TimeProfile Maximum possible installed rate of the technology in each period.

rate_max_add::TimeProfile Maximum rate addition in a strategic period.

rate_min_add::TimeProfile Minimum rate addition in a strategic period.

capex_stor::TimeProfile Capital expenditure, here investment costs of the technology storage volume in each strategic period.

stor_max_inst::TimeProfile Maximum possible installed storage volume of the technology in each strategic period.

stor_max_add::TimeProfile Maximum storage volume addition in one period from the previous.

stor_min_add::TimeProfile Minimum storage volume addition in one period from the previous.

inv_mode::Investment = ContinuousInvestment() Type of the investment: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

rate_start::Union{Real, Nothing} = nothing Starting rate in first period. If nothing is given, it is set by start_cap() to the capacity rate_cap of the node in the first strategic period.

stor_start::Union{Real, Nothing} = nothing Starting storage volume in first period. If nothing is provided, it is set by start_cap() to the capacity stor_cap of the node in the first strategic period.

rate_increment::TimeProfile = FixedProfile(0) Rate increment used in the case of DiscreteInvestment

stor_increment::TimeProfile = FixedProfile(0) Storage volume increment used in the case of DiscreteInvestment

life_mode::LifetimeMode = UnlimitedLife() Type of handling of the lifetime: UnlimitedLife, StudyLife, PeriodLife, or RollingLife

lifetime::TimeProfile = FixedProfile(0) Duration/lifetime of the technology invested in each period.

TransInvData

Similarly as for InvData, this type defines additional parameters necessary for handling the investment in transmission between geographical areas. This type is used in combination with EnergyModelsGeography to add investments in transmission.

There are in general not too many changes compared to InvData. The individual fields have however different names:

capex_cap is called capex_trans,

cap_max_inst is called trans_max_inst,

cap_max_add is called trans_max_add,

cap_min_add is called trans_min_add,

cap_start is called trans_start, and

cap_increment is called trans_increment.

In addition, the following field is added:

capex_trans_offset::TimeProfile: The offset is a special parameter only required for the SemiContinuousOffsetInvestment mode. The offset can be best described with the equation $y = \texttt{capex\_trans} \times x + \texttt{capex\_trans\_offset}$ where $x$ corresponds to the invested capacity and $y$ to the total capital cost.

EnergyModelsInvestments.TransInvData — Type

Extra data for investing in transmission.

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses the macro Base.@kwdef to use keyword arguments and default values. Hence, the name of the parameters have to be specified.

Fields

capex_trans::TimeProfile Capital expenditure for the transmission capacity, here investment costs of the transmission in each period.

trans_max_inst::TimeProfile Maximum possible installed transmission capacity in each period.

trans_max_add::TimeProfile Maximum transmission capacity addition in one period from the previous.

trans_min_add::TimeProfile Minimum transmission capacity addition in one period from the previous.

inv_mode::Investment = ContinuousInvestment() Type of the investment: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

trans_start::Union{Real, Nothing} = nothing Starting transmission capacity in first period. If nothing is given, it is set by getstartcap() to the capacity trans_cap of the transmission.

trans_increment::TimeProfile = FixedProfile(0) Transmission capacity increment used in the case of DiscreteInvestment

Investment modes

Different investment modes are available to help represent different situations. The investment mode is included in the field inv_mode in InvData, InvDataStorage, and TransInvData. The investment mode determines which other parameters are relevant in the investment and how these are treated.

Investment

Investment is the abstract supertype for all investment modes. It is used to allow for a simple extension of the potential investment modes. It is also possible for the user to define new investment modes without major changes to the core structure through specifying a new subtype of Investment.

EnergyModelsInvestments.Investment — Type

Investment type traits for nodes.

The investment type corresponds to the chosen investment mode.

ContinuousInvestment

ContinuousInvestment is the default investment option for all investments, if no alternative is chosen. Continuous investments implies that you can invest in any capacity specified between cap_min_add and cap_max_add. This implies as well that, if cap_min_add is specified, it is necessary to invest in every strategic period in at least this capacity. This approach is the standard approach in large energy system models as it avoids binary variables. However, it can lead to, e.g., investments into a 10~MW nuclear power plant.

Fields without a meaning in ContinuousInvestment:

cap_increment

(trans_capex_offset if using TransInvData)

Warning

Defining cap_min_add::TimeProfile for this mode of investment will lead to a forced investment of at least cap_min_add in each period.

EnergyModelsInvestments.ContinuousInvestment — Type

Continuous investment between a minimum and a maximum value.

BinaryInvestment

BinaryInvestment implies that one can choose to invest in the specified capacity (field cap of the node) in the given strategic period, or not. The capacity of the investment cannot be adjusted by the optimization. This implies that the meaning of the capacity of a Node, cap, is redefined. Hence, it is important to specify cap_start to avoid issues in the first strategic period.

Fields without a meaning in BinaryInvestment:

cap_min_add

cap_max_add

cap_increment

(trans_capex_offset if using TransInvData)

Warning

This investment type leads to the addition of binary variables. The number of binary variables is equal to the number of strategic periods times the number of Nodes with the BinaryInvestment mode.

EnergyModelsInvestments.BinaryInvestment — Type

Binary investment in given capacity with binary variables. Requires specification of cap_start in InvData for proper analyses.

DiscreteInvestment

DiscreteInvestment allow for only a discrete increase in the capacity. This increase is specified through the field cap_increment. Hence, it can be also dependent on the strategic period.

DiscreteInvestment can for example be used to represent investment in modular technologies that can be scaled by adding several modules together. In addition, it is beneficial to include for technologies that experience significant economy of scale. In this situation, several instances with different cap_increment and capex_cap can be used

Fields without a meaning in DiscreteInvestment:

cap_min_add

cap_max_add

(trans_capex_offset if using TransInvData)

Note

This investment type leads to the addition of integer variables. The number of integer variables is equal to the number of strategic periods times the number of Nodes with the DiscreteInvestment mode.

EnergyModelsInvestments.DiscreteInvestment — Type

Investment in fixed increments with integer variables.

SemiContiInvestment

SemiContiInvestment is an abstract type used for two investment modes:

SemiContinuousInvestment and

SemiContinuousOffsetInvestment.

These investment modes are similar with respect to how you can increase the capacity. They differ however on how the overall cost is calculated. Both investment modes are in general similar to ContinuousInvestment, but the investment is either 0 or between a minimum and maximum value. This means you can define the field cap_min_add::TimeProfile without forcing investment in the technology. Instead, the value determines that if the model decides to invest, then it has to at leas invest in the value provided through `capminadd`. This can be also described as:

$x = 0 \lor \texttt{cap\_min\_add} \leq x \leq \texttt{cap\_max\_add}$

with $x$ corresponding to the invested capacity

Fields without a meaning in SemiContiInvestment:

cap_increment

EnergyModelsInvestments.SemiContiInvestment — Type

Semi-continuous investment, either zero or between a minimum and a maximum value, involves a binary variable.

Note

These investment modes leads to the addition of binary variables. The number of binary variables is equal to the number of strategic periods times the number of Nodes with the SemiContinuousInvestment and SemiContinuousOffsetInvestment mode.

SemiContinuousInvestment

The cost function in SemiContinuousInvestment is calculated in the same way as in ContinuousInvestment. The total contribution of invested capacity $x$ to the objective function $y$ is given through the equation

$y = \texttt{capex\_cap} \times x$.

Fields without a meaning in SemiContinuousInvestment:

(trans_capex_offset if using TransInvData)

EnergyModelsInvestments.SemiContinuousInvestment — Type

Semi-continuous investment where the cost is going through the origin.

SemiContinuousOffsetInvestment

SemiContinuousOffsetInvestment is a type of investment similar to SemiContinuousInvestment and implemented for investments in transmission infrastructure. It does differ with respect to how the costs are calculated. A SemiContinuousOffsetInvestment has an offset in the cost implemented through the the field Capex_trans_offset. This offset corresponds to the theoretical cost at an invested capacity of 0.

While SemiContinuousInvestmentutilizes the same relative cost, even if a lower limit is specified, SemiContinuousOffsetInvestment allows for the specification of an offset in the cost through the field trans_capex_offset. This offset is an absolute cost. It corresponds to the theoretical cost at an invested capacity of 0. This changes the contribution to the cost function from

$y = \texttt{capex\_cap} \times x$

to

$y = \texttt{capex\_trans} \times x + \texttt{capex\_trans\_offset}$

where $x$ corresponds to the invested capacity and $y$ to the total capital cost.

SemiContinuousOffsetInvestment is currently only implemented for investments in TransmissionModes as its implementation would interact with the lifetime calculations.

EnergyModelsInvestments.SemiContinuousOffsetInvestment — Type

Semi-continuous investment where the cost has an additional offset

FixedInvestment

FixedInvestment is a type of investment where an investment in the given capacity is forced. The capacity used is provided through the fields cap, cap_stor, cap_rate, and trans_cap. It allows to account for the investment cost of known investments. In practice, there is however not too much use in including the fixed investment, except if one is interested in the values of the dual variables.

The fields cap_min_add, cap_max_add, and cap_increment do not have a meaning when using FixedInvestment.

Fields without a meaning in DiscreteInvestment:

cap_min_add

cap_max_add

cap_increment

(trans_capex_offset if using TransInvData)

EnergyModelsInvestments.FixedInvestment — Type

Forced investment in given capacity.

LifetimeMode

EnergyModelsInvestments allows for differing descriptions for the lifetime of a technology. A key problem is when the lifetime of a technology is not equal to the duration of strategic periods. To this end, several ways to define the lifetime of a technology are available in the package and presented below.

It is also possible for the user to define new LifetimeModes. In practice, this requires only the introduction of a new subtype to LifetimeMode as well as a single function.

EnergyModelsInvestments.LifetimeMode — Type

Abstract lifetime mode type.

UnlimitedLife

This LifetimeMode is used when the lifetime of a Node is not limited. No reinvestment is considered by the optimization and there is also ne salvage value (or rest value) at the end of the last strategic period. Hence, the costs are the same, independent of if the investments in the Node are happening in the first strategic period (and the technology is used for, e.g, 25 years) or the last strategic period (with a usage of, e.g., 5 years) when excluding discounting effects.

UnlimitedLife is the default lifetime mode, if no other mode is specified.

The field Lifetime does not have a meaning when using UnlimitedLife.

EnergyModelsInvestments.UnlimitedLife — Type

The investment's life is not limited. The investment costs do not consider any reinvestment or rest value.

StudyLife

StudyLife includes the technology for the full investigated horizon. If the Lifetime is shorter than the remaining horizon, reinvestments are considered. These reinvestments are included in the costs of the investment strategic period, but discounted to their actual value.

As an example, consider investments with a lifetime of 20 years in 2030, while the study horizon ends in 2055. In this situation, reinvestments are required in 2050 to allow for operation in the last 5 years. The CAPEX are then correspondingly adjusted to account for both discounted reinvestments and final value in 2055.

EnergyModelsInvestments.StudyLife — Type

The investment lasts for the whole study period with adequate reinvestments at the end of the lifetime and considering the rest value.

PeriodLife

PeriodLife is used to define that the investment is only lasting for the strategic period in which it happens. Additional year of lifetime are counted as a rest value. Reinvestment inside the strategic periods are also considered in case the lifetime is shorter than the length of the strategic period.

EnergyModelsInvestments.PeriodLife — Type

The investment is considered to last only for the strategic period. The excess lifetime is considered in the rest value. If the lifetime is lower than the length of the period, reinvestment is considered as well.

RollingLife

RollingLife corresponds to the classical roll-over of investments from one strategic period to the next until the end of life is reached. In general, three different cases can be differentiated:

The lifetime is shorter than the duration of the strategic period. In this situation, a PeriodLife is assumed.

The lifetime equals the duration of the strategic period. In this situation, the capacity is retired at the end of the strategic period

The lifetime is longer than the duration of the strategic period. This leaves however a problem if the lifetime does fall in-between two strategic periods, as it would be the case for a lifetime of, e.g., 8 years and two strategic periods of, e.g, 5 years. In this case, the technology would only be available for the first 3 years of the second strategic period leaving the question on how to handle this situation. EnergyModelsInvestments retires the technology at the last full strategic period and calculates the remaining value for the technology.

EnergyModelsInvestments.RollingLife — Type

The investment is rolling to the next strategic periods and it is retired at the end of its lifetime or the end of the previous strategic period if its lifetime ends between two periods.